YH-602: fix DocsPage
