+++ src/pages/landing/DocsPage/ui/DocsPage.module.css
@@ -1,0 +1,7 @@
+@media screen and (width <= 576px) {
+  .container {
+    padding: 0 15px;
+  }
+}
+  
+
