+++ src/widgets/Docs/DocsBlock/index.ts
@@ -1,3 +1,3 @@
-export { Docs } from './ui/Docs/Docs';
+export { DocsList } from './ui/Docs/DocsList';
 export { DocItem } from './ui/DocItem/DocItem';
-export type { DockProps } from './model/types/types';
+export type { DockItem } from './model/types/types';
